refactor: introduce BaseComBat and use it as base class for combat-base methods
